\name{capacity}
\docType{data}
\alias{capacity}
\title{capacity}
\description{Cross-national panel data on democracy and state capacity.}
\format{dataframe}
\keyword{datasets}
\references{
  The democracy measure (demo) comes from Boix, Miller, and Rosato (2013). The state capacity measure (Capacity) comes from Hanson and Sigman (2013.)

  Reference: Erik H. Wang and Yiqing Xu (2018). "Awakening Leviathan: the Effect of Democracy on State Capacity, 1960-2009." Research and Politics, forthcoming. Available at: \url{http://journals.sagepub.com/doi/full/10.1177/2053168018772398}.
}

