\name{panelView-package}
\alias{panelView-package}
\docType{package}
\title{Panel Data Visualizations}
\description{Visualizes panel data with (dichotomous) treatments}

\details{
  panelView has two main functionalities: (1) it visualizes the treatment and missing-value statuses of each observation in a panel/time-series-cross-sectional (TSCS) dataset; and (2) it plots the outcome variable (either continuous or discrete) in a time-series fashion.

  For more details, see \url{https://yiqingxu.org/packages/panelView/panelView.html}.
}

\author{
  Licheng Liu <liulch.16@sem.tsinghua.edu.cn>, Tsinghua University  

  Yiqing Xu <yiqingxu@ucsd.edu>, University of California, San Diego
}
