\name{simdata}
\docType{data}
\alias{simdata}
\title{A simulated dataset}
\description{A simulated panel dataset with a discrete outcome variable.}
\format{dataframe}
\keyword{datasets}
\references{
Hongyu Mou, Licheng Liu and Yiqing Xu (2023). "Panel Data Visualization in {R} ({panelView}) and {Stata} ({panelview})." Journal of Statistical Software, 107(7), pp. 1--20. <doi:10.18637/jss.v107.i07>
}

