\name{capacity}
\docType{data}
\alias{capacity}
\title{capacity}
\description{Cross-national panel data on democracy and state capacity.}
\format{dataframe}
\keyword{datasets}
\references{
  The democracy measure (demo) comes from Boix, Miller, and Rosato (2013). The state capacity measure (Capacity) comes from Hanson and Sigman (2013.)

  Reference: Erik H. Wang and Yiqing Xu (2018). "Awakening Leviathan: the Effect of Democracy on State Capacity, 1960-2009." Research and Politics, Vol. 5, Iss. 2, April-June 2018, pp. 1-7.
}

