\name{panelview-package}
\alias{panelview-package}
\docType{package}
\title{Panel Data Visualizations}
\description{Visualizes panel data with (dichotomous) treatments}

\details{
  panelview has three main functionalities:

(1) it visualizes the treatment and missing-value statuses of each observation in a panel dataset;

(2) it plots the outcome variable (either continuous or discrete) in a time-series fashion;

(3) it visualizes the relationships between the outcome and treatment variable individually or in an aggregate fashion.

  For more details, see \url{https://yiqingxu.org/packages/panelview/panelview.html}.
}

\author{
  Hongyu Mou <hongyumou5@gmail.com>, Peking Univeristy
  
  Licheng Liu <liulch.16@sem.tsinghua.edu.cn>, Tsinghua University  

  Yiqing Xu <yiqingxu@stanford.edu>, Stanford University
}
