\name{panelView-package}
\alias{panelView-package}
\docType{package}
\title{Panel Data Visualizations}
\description{Visualizes panel data}

\details{
  panelView has three main functionalities:

(1) it plots treatment status and missing data in a panel dataset;

(2) it plots an outcome variable (either continuous or discrete) in a time-series fashion;

(3) it visualizes the bivariate relationships between an outcome variable and a treatment variable by unit or in aggregate.
}

\author{
  Hongyu Mou <hongyumou@g.ucla.edu>, University of California, Los Angeles

  Licheng Liu <liulch@mit.edu>, Massachusetts Institute of Technology

  Yiqing Xu <yiqingxu@stanford.edu>, Stanford University
}
